Advance RinOS WebGL and WASI integration

+++ Libraries/LibWeb/WebAssembly/Memory.h
@@ -52,7 +52,7 @@ private:
 
     static void refresh_the_memory_buffer(JS::VM&, JS::Realm&, Wasm::MemoryAddress);
     static GC::Ref<JS::ArrayBuffer> create_a_fixed_length_memory_buffer(JS::VM&, JS::Realm&, Wasm::MemoryAddress, Shared shared);
-    static JS::ThrowCompletionOr<GC::Ref<JS::ArrayBuffer>> create_a_resizable_memory_buffer(JS::VM&, JS::Realm&, Wasm::MemoryAddress, Shared shared, size_t max_size);
+    static JS::ThrowCompletionOr<GC::Ref<JS::ArrayBuffer>> create_a_resizable_memory_buffer(JS::VM&, JS::Realm&, Wasm::MemoryAddress, Shared shared, u64 max_size);
 
     Wasm::MemoryAddress m_address;
     Shared m_shared { Shared::No };
